Implemented WebView Auto Refresh using Network Observer
A coroutine scope is started in the main activity, in oder to prevent work leaks. onDestroy, all coroutines are cancelled.

The webview client takes the coroutine scope, again to avoid work leaks, and utilizes it to check if the user is offline when the webview client receives an error

Once the webview client receives an error, an Event indicating the user is offline is posted.

The webview client is subscribed to the NetworkStatusEvent in order to know the network status. Once the device is offline, the network watcher is started (a broadcast receiver waits until the device is back online). Once the device is back online, the network watcher is stopped and the webview is refreshed

It is important to note that the webview client is not automatically registered to the Event. This is because we want to avoid leaks. Therefore, whenever it is manually registered, it should be manually unregistered.