Implemented Auto Refresh for WebView (#77)
* Updated Libs using refreshVersions

Updated Libs and changed strings in implementations to variables in Libs object

* Added GreenRobot EventBus to Dependencies

* Added Kotlin Coroutines Dependency

Updated Kotlin Version to v1.3.71 since Kotlin Coroutines v1.3.7 are only supported with Kotlin v1.3.7+

* Enabled MutliDex for DEV Android App

Added multidex following android docs, https://developer.android.com/studio/build/multidex

* Added NetworkStatusEvent and NetworkStatus for EventBus

NetworkStatusEvent will be used with EventBus, it will carry the NetworkStatus to Listeners or Subscribers to that Event

Once the NetworkStatusEvent is received, certain actions will be taken, depending on if the device is online or offline.

* Added NetworkUtils to Check if Device is Online/Offline

Checking if user is online is done through opening a socket with 8.8.8.8 (google dns). This online check is according to the stackoverflow answer linked below. It is faster than other options.

https://stackoverflow.com/a/27312494/5512292

* Added NetworkWatcher BroadcastReceiver to Know When Device is Back Online

This is used in order to check if the device is back online. Once the device's connectivity changes, ie is connected to internet after being offline, it posts an Event indicating the device is back online.

* Implemented WebView Auto Refresh using Network Observer

A coroutine scope is started in the main activity, in oder to prevent work leaks. onDestroy, all coroutines are cancelled.

The webview client takes the coroutine scope, again to avoid work leaks, and utilizes it to check if the user is offline when the webview client receives an error

Once the webview client receives an error, an Event indicating the user is offline is posted.

The webview client is subscribed to the NetworkStatusEvent in order to know the network status. Once the device is offline, the network watcher is started (a broadcast receiver waits until the device is back online). Once the device is back online, the network watcher is stopped and the webview is refreshed

It is important to note that the webview client is not automatically registered to the Event. This is because we want to avoid leaks. Therefore, whenever it is manually registered, it should be manually unregistered.

* Changed Online Check Hostname to dev.to

Uses hostname instead of url, so created BuildConfig.baseHostname instead of trimming the 'https://' and '/' segments of the BuildConfig.baseUrl. This gives for easier changes later on in the future.

Also, @fdoxyz requested that change in order to reconnect to the website once it is up again, if it was down. Mentioned in pull request #77.

* Cleared Network Watcher Variable to Null on Unregistration in WebViewClient

Previously, an exception was being thrown due to unregistering a network watcher and then trying to unregister it again later on. Now, since it is being cleared back to null, it is no longer being unregistered again if it was previously unregistered

* Fixed Multiple Re-Registrations for Network Watcher

onReceiveError was being called multiple times, this was resulting in multiple events being posted indicating that the device is offline. Accordingly, it was important to stop re-registering the networkWatcher again and again whenever the events were captured. However, it was previously done in case the coroutineScope did change. However, if the coroutineScope changes, the activity has probably died and been recreated. Therefore, the webview would have already been destroyed and recreated with the new scope. Thus, there is no need to re-register a network watcher if it has already been registered

Co-authored-by: Fernando Valverde <fdov88@gmail.com>